@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.3",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "apollo-boost": "^0.4.7",
-    "bootswatch": "^4.4.1",
-    "graphql": "^14.6.0",
-    "graphql-tag": "^2.10.3",
-    "react": "^16.13.1",
-    "react-apollo": "^3.1.3",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1"
+    "@apollo/react-hooks": "^4.0.0",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "apollo-boost": "^0.4.9",
+    "bootswatch": "^5.1.3",
+    "graphql": "^16.2.0",
+    "graphql-tag": "^2.12.6",
+    "react": "^17.0.2",
+    "react-apollo": "^3.1.5",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
