@@ -1,8 +1,9 @@
-import mongoose from 'mongoose'
+import mongoose from "mongoose";
 
-mongoose.connect('mongodb://localhost/testdb', {
-  useNewUrlParser: true,
-  useUnifiedTopology: true
-})
-  .then(db => console.log('Db is connected'))
-  .catch(err => console.log(err));
+mongoose
+  .connect("mongodb://localhost/testdb", {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  })
+  .then((db) => console.log("Db is connected"))
+  .catch((err) => console.log(err));
