@@ -4,7 +4,7 @@ const Mutation = {
   createMessage: async (_, { title, content, author }) => {
     const newMessage = new Message({ title, content, author });
     return await newMessage.save();
-  }
+  },
 };
 
 export default Mutation;
