@@ -6,7 +6,7 @@ const Query = {
   },
   messages: async () => {
     return await Message.find();
-  }
+  },
 };
 
 export default Query;
