@@ -1,7 +1,7 @@
-import Query from './Query'
-import Mutation from './Mutation'
+import Query from "./Query";
+import Mutation from "./Mutation";
 
 export default {
   Query,
-  Mutation
-}
+  Mutation,
+};
