@@ -1,6 +1,6 @@
-import {server} from './server'
-import './database';
+import { server } from "./server";
+import "./database";
 
-server.start({port: 3100}, ({port}) => {
-  console.log('Server on port', port);
+server.start({ port: 3100 }, ({ port }) => {
+  console.log("Server on port", port);
 });
