@@ -3,16 +3,16 @@ import { Schema, model } from "mongoose";
 const messageSchema = new Schema({
   title: {
     type: String,
-    required: true
+    required: true,
   },
   content: {
     type: String,
-    required: true
+    required: true,
   },
   author: {
     type: String,
-    required: true
-  }
+    required: true,
+  },
 });
 
 export default model("Message", messageSchema);
