@@ -1,6 +1,8 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import { useMutation } from "@apollo/react-hooks";
 import { gql } from "apollo-boost";
+import { useNavigate } from "react-router-dom";
+import { GET_MESSAGES } from "./MessageList";
 
 const CREATE_MESSAGE = gql`
   mutation CreateMessage($title: String!, $content: String!, $author: String!) {
@@ -15,7 +17,12 @@ export default function MessageForm(props) {
   const [title, setTitle] = useState("");
   const [author, setAuthor] = useState("");
 
-  const [createMessage] = useMutation(CREATE_MESSAGE);
+  const [createMessage] = useMutation(CREATE_MESSAGE, {
+    // fetch the list of messages again for update
+    refetchQueries: [{ query: GET_MESSAGES }],
+  });
+
+  const navigate = useNavigate();
 
   return (
     <div className="row">
@@ -23,35 +30,35 @@ export default function MessageForm(props) {
         <div className="card">
           <div className="card-body">
             <form
-              onSubmit={async e => {
+              onSubmit={async (e) => {
                 e.preventDefault();
                 await createMessage({ variables: { title, author, content } });
-                window.location.href = "/";
+                navigate("/");
               }}
             >
-              <div className="form-group">
+              <div className="my-2">
                 <input
                   type="text"
                   placeholder="Author"
                   value={author}
-                  onChange={e => setAuthor(e.target.value)}
+                  onChange={(e) => setAuthor(e.target.value)}
                   className="form-control"
                 />
               </div>
-              <div className="form-group">
+              <div className="my-2">
                 <input
                   type="text"
                   placeholder="Write a Title..."
                   value={title}
-                  onChange={e => setTitle(e.target.value)}
+                  onChange={(e) => setTitle(e.target.value)}
                   className="form-control"
                 />
               </div>
-              <div className="form-group">
+              <div className="my-2">
                 <textarea
                   rows="2"
                   placeholder="Content..."
-                  onChange={e => setContent(e.target.value)}
+                  onChange={(e) => setContent(e.target.value)}
                   value={content}
                   className="form-control"
                 ></textarea>
