@@ -2,7 +2,7 @@ import React from "react";
 import {Link} from 'react-router-dom'
 
 export const Navigation = () => (
-  <nav className="navbar navbar-expand-lg navbar-light bg-light">
+  <nav className="navbar navbar-expand-lg navbar-light bg-light p-4">
     <Link className="navbar-brand" to="/">
       ReactAndGraphQL
     </Link>
@@ -18,7 +18,7 @@ export const Navigation = () => (
       <span className="navbar-toggler-icon" />
     </button>
     <div className="collapse navbar-collapse" id="navbarNav">
-      <ul className="navbar-nav">
+      <ul className="navbar-nav ms-auto">
         <li className="nav-item">
           <Link className="nav-link" to="/">
             Messages
